requirements.txt
psycopg2-binary 
pandas 
configparser
